feat: Add updateComplement function to update course complement
This commit adds the updateComplement function, which is responsible for updating the course complement. The function takes in the options object containing the complement data and calls the updateCoursComplement function from the complement usecases. This feature enhances the functionality of the application by allowing users to update the course complement easily.