feat: Add functionality to add student to class
This commit adds the functionality to add a student to a class in the StudentsEvaluationTableView component. It introduces the AddStudentForm component, which allows users to input the necessary information for adding a student. When the "Add Student" button is clicked, the form values are retrieved and the student is added to the class using the addStudent function. If the student is successfully added, the form is reset and a success toast message is displayed. Otherwise, an error toast message is displayed. This enhancement improves the codebase by providing an easy way for users to add students to classes.