refactor: Update SpacesComplementServerLayer to use LayoutWithProps isEmpty prop
This commit updates the SpacesComplementServerLayer component to use the isEmpty prop of the LayoutWithProps component. By setting the isEmpty prop to true, the layout will render a "nothing to show" message when there is no complement content available. This improvement enhances the user experience by providing clear feedback when there is no content to display in the spaces complement module.