refactor: Add index on userId in convex schema
This commit adds an index on the "userId" column in the convex schema. The index is named "by_userId" and improves the performance of queries that involve filtering or sorting by the "userId" field. This change enhances the efficiency of the database operations and improves the overall performance of the application.