feat: Add publish flag to updateCours mutation
This commit adds a new "publish" flag to the updateCours mutation in the ConvexDatabase. The flag allows for controlling the visibility of the updated cours. If the flag is set to true, the cours will be published and visible to users. If the flag is set to false, the cours will be unpublished and not visible to users. This enhancement improves the flexibility of the application and provides better control over the visibility of cours.