feat: Add publish flag to createCours mutation
This commit adds a new "publish" flag to the createCours mutation in the ConvexDatabase. The flag allows for controlling the visibility of the created cours. If the flag is not provided, it defaults to false. This enhancement improves the flexibility of the application and provides better control over the visibility of cours.