refactor: Delete complement mutation in ConvexDatabase
This commit adds a new deleteComplement mutation to the ConvexDatabase. The mutation allows for deleting a complement by its ID. The mutation first queries the complement with the provided ID and then deletes it from the database. If the complement is found and successfully deleted, the mutation returns true. Otherwise, it returns false. This enhancement improves the functionality of the application by providing the ability to delete complements from the database.