refactor: Update SequenceServerLayer to include "All Courses" link
This commit updates the SequenceServerLayer component in the src/app/sequences/[slug]/SequenceServerLayer.tsx file. It adds a new navigation item titled "All Courses" with a link to `/cours/all/${props.slug}`. This change enhances the navigation options for users accessing the sequence page, allowing them to easily navigate to all courses related to the sequence.