refactor: Delete sequence and associated courses
This commit refactors the CoursRepository and CoursUsecases to add the functionality for deleting a sequence and its associated courses. It introduces a new method called deleteSequence that takes in the sequenceId and type parameters to identify the sequence to be deleted. The method first retrieves all the courses associated with the sequence and deletes them individually. Then, it deletes the sequence itself from the database. This enhancement improves the data management capabilities of the application.