refactor: Add useDeleteSequence hook for deleting sequences
This commit adds a new hook called useDeleteSequence in the useDeleteSequence.ts file. The hook utilizes the useMutation function from the @tanstack/react-query library to handle the deletion of sequences. It takes in the sequenceId and type parameters to identify the sequence to be deleted. Upon successful deletion, a success toast message is displayed and the page is reloaded. This enhancement improves the data management capabilities of the application.