refactor: Add ClasseSequencesServerLayer for fetching classe sequences
This commit adds the ClasseSequencesServerLayer component to the src/app/classes/sequences/[slug]/ClasseSequencesServerLayer.tsx file. The component is responsible for fetching and displaying the sequences for a specific classe. It makes use of the coursUsecases.getAllSequences and coursUsecases.getClasseSequences functions to retrieve the sequences data. This enhancement improves the data retrieval and presentation capabilities of the application.