feat: Add deleteEvaluationBase method to ConvexDatabase and IDatabase
This commit adds the deleteEvaluationBase method to the ConvexDatabase class and the IDatabase interface. The method allows for the deletion of evaluation bases. It takes in the evaluationId as a parameter and uses the fetchMutation function to send a request to delete the evaluation base. If successful, it returns a Promise with a void value. If an error occurs, it returns a Promise with a Failure object containing the error details. This enhancement improves the data management capabilities of the application.