refactor: Update UpdateStudentGradeForm component
This commit refactors the UpdateStudentGradeForm component to improve code readability and maintainability. It removes unused imports and code related to the useGetEvaluationCompoundList hook, removes unnecessary state and useEffect, updates dependencies of the useEffect that sets evaluation base criterias, removes the useEffect that updates local table data based on the compound list, and updates props passed to the TenPointsCriteriaForm component. These changes simplify the component's logic and remove unnecessary code.