refactor: Add refetchStudents function to AddStudentForm and CSVReader components
This commit adds the refetchStudents function to the AddStudentForm and CSVReader components in the StudentsEvaluationTableView.tsx file. The refetchStudents function is called after a successful student addition or CSV file import, triggering a refetch of the student data to update the table view. This enhancement improves the functionality of the components and ensures that the table view is always up to date with the latest student data.