refactor: Update staleTime for queries to improve data freshness
This commit updates the staleTime property for queries in the QueryProvider.tsx file. The previous value of 10 minutes has been changed to 1 minute, ensuring that the data fetched from the server is more up to date. This change improves the freshness of the data displayed to the user, providing a better user experience.