```text feat: Add useDeleteClasse hook for deleting a class
This commit adds the useDeleteClasse hook to the useDeleteClasse.ts file. The hook provides a convenient way to delete a class by making use of the useMutation hook from the @tanstack/react-query library. The hook takes in the class ID as a parameter and performs a mutation to delete the class using the classeUsecases.deleteClasse function. This addition enhances the functionality of the application by providing a reusable and simplified way to delete classes.