refactor: Update grade type values in grade_type.ts
This commit updates the grade type values in the grade_type.ts file. The previous values "Excellent/Good/Satisfactory/Needs Improvement" have been changed to "Expertise/Proficiency/To be developed/To be acquired". This change improves the clarity and accuracy of the grade types, aligning them with the desired terminology.